==> 02 - Busca dados SQL para CSV.ipynb <==
## Etapa 2: Busca dos tweets coletados

Após a coleta dos Tweets, vamos trazer esses dados para o formato CSV e assim trabalharmos localmente. Para isso vamos usar a biblioteca Pandas. Neste jupyter notebook não faremos nenhum tratamento do texto, deixaremos isso para um segundo momento.

import pypyodbc as podbc
import pandas as pd

def conexao_sql():
    conexao = podbc.connect("Driver=ODBC Driver 17 for SQL Server;Server=posml.database.windows.net,1433;Database=tweets;uid=cloudadmin;pwd=senha")
    return conexao

# Retornando a conexão SQL
conn = conexao_sql()

# Testando a consutal SQL
sql = "SELECT * FROM Tweets "

# Transformando a consulta em um dataset pandas
df = pd.read_sql(sql, conn)
df.head()

# verificando se todas as empresas monitoradas tem registros
df['in_reply_to_screen_name'].value_counts()

# Salvando o dataset em CSV
df.to_csv('dataset.csv')

print('Arquivo Gravado com Sucesso')

# Validando a importação do CSV
dataset = pd.read_csv('dataset.csv', index_col=0)
dataset.head()
