==> 03 - Google Play Scraper.ipynb <==
## Etapa 3: Coleta do dataset rotulado

Para conseguirmos treinar um modelo de previsão de sentimentos, precisamos de uma base rotulada, mas conseguir uma base em português, de tweets e dentro do contexto, não é uma tarefa facil! Assim decidimos criar a nossa propria base rotulada, em português e de certa forma contextualizada. Para isso utilizamos os comentarios dos Apps das empresas monitoradas na Google Store, onde tomamos a pressisa de que notas 5 estrelas são comentários positivos e nota 1 estrela comentários negativos. Não utilizamos as notas 2, 3 e 4.

!pip install google-play-scraper -q

from google_play_scraper import app
from google_play_scraper import Sort, reviews_all
import pandas as pd
import json
import datetime

#Setando a url do app das empresas  

nubank = 'com.nu.production'
banco_inter = 'br.com.intermedium'
c6bank = 'com.c6bank.app'
timeneon = 'br.com.neon'
Bradesco = 'com.bradesco'
santander_br = 'com.santander.app'  
BancodoBrasil = 'br.com.bb.android'   
itau = 'com.itau'
 
NetflixBrasil = 'com.netflix.mediaclient'
PrimeVideoBR = 'com.amazon.avod.thirdpartyclient'
globoplay = 'com.globo.globotv'
DisneyPlusBR = 'com.disney.disneyplus'

submarino = 'com.b2w.submarino'
americanascom = 'com.b2w.americanas'
pontofrio = 'com.sonyericsson.pontofrio'
amazonBR = 'com.amazon.mShop.android.shopping'
CasasBahia = 'com.novapontocom.casasbahia'

CocaCola_Br = ''
PepsiBr = ''
brkindle = ''

Lista_Bancos = [nubank, banco_inter, c6bank, timeneon, Bradesco, santander_br, BancodoBrasil, itau]
Lista_Streamers = [NetflixBrasil, PrimeVideoBR, globoplay, DisneyPlusBR]
Lista_marketplace = [submarino, americanascom, pontofrio, amazonBR, CasasBahia]
#Lista_outros = [CocaCola_Br, PepsiBr, brkindle]

def busca_review_unico(app):
    
    positivos = reviews_all(
    app,
    sleep_milliseconds=0, # defaults to 0
    lang='pt', # defaults to 'en'
    country='br', # defaults to 'us'
    sort=Sort.MOST_RELEVANT, # defaults to Sort.MOST_RELEVANT
    filter_score_with=5 # defaults to None(means all score)
    )
    
    negativos = reviews_all(
    app,
    sleep_milliseconds=0, # defaults to 0
    lang='pt', # defaults to 'en'
    country='br', # defaults to 'us'
    sort=Sort.MOST_RELEVANT, # defaults to Sort.MOST_RELEVANT
    filter_score_with=1 # defaults to None(means all score)
    )
    print('APP : ', app)
    print('Reviews Positivos : ',len(positivos))
    print('Reviews Negativos : ',len(negativos))
    print('-----------------------------------------')
    return positivos, negativos
    

def busca_review_lista(lista_app):
    rev_positivos, rev_negativos = [],[]
    for apps in  lista_app:
        positivos, negativos = busca_review_unico(apps)
        rev_positivos += positivos
        rev_negativos += negativos
    
    print('-----------------------------------------')    
    print('Total Reviews Positivos : ',len(rev_positivos))
    print('Total Reviews Negativos : ',len(rev_negativos))
    print('-----------------------------------------')
    return rev_positivos, rev_negativos

def defaultconverter(o):
    if isinstance(o, datetime.datetime):
        return o.__str__()

def converte_jason_para_dataframe(lista_positiva, lista_negativa):
    rev_positivos_json = json.dumps(lista_positiva, default = defaultconverter)
    rev_negativos_json = json.dumps(lista_negativa, default = defaultconverter)
    
    df_positivos = pd.read_json(rev_positivos_json)
    df_negativos = pd.read_json(rev_negativos_json)
    
    df_positivos=df_positivos[(['content','score'])]
    df_negativos=df_negativos[(['content','score'])]
    
    df_ = pd.concat([df_positivos, df_negativos], ignore_index=True)
    
    return df_

# Capturando os dados da lista de bancos
rev_positivos, rev_negativos = busca_review_lista(Lista_Bancos)

# Armazenando em um pandas dataset
df_bancos = converte_jason_para_dataframe(rev_positivos, rev_negativos)
df_bancos

# Exportando pra um CSV
df_bancos.to_csv('reviews_bancos.csv')

# Capturando os dados da lista de Streamers
rev_positivos_stream, rev_negativos_stream = busca_review_lista(Lista_Streamers)

# Armazenando em um pandas dataset
df_streamers = converte_jason_para_dataframe(rev_positivos_stream, rev_negativos_stream)
df_streamers

# Exportando pra um CSV
df_streamers.to_csv('reviews_streamers.csv')

# Capturando os dados da lista de marketplace
rev_positivos_marketplace, rev_negativos_marketplace = busca_review_lista(Lista_marketplace)

# Armazenando em um pandas dataset
df_marketplace = converte_jason_para_dataframe(rev_positivos_marketplace, rev_negativos_marketplace)
df_marketplace

# Exportando pra um CSV
df_streamers.to_csv('reviews_marketplace.csv')

# Unindo todos os datasets
df_geral = pd.concat([df_bancos, df_streamers, df_marketplace], ignore_index=True)
df_geral

# Verificando se o dataset ficou Balanceado
df_geral.score.value_counts()

# removendo comentarios para balancear o dataset
df_geral = df_geral[:78000]

# Exportando o dataset Unificado para um CSV
df_geral.to_csv('reviews_geral.csv')
