==> 04 - Limpeza dos datasets.ipynb <==
## Etapa 4: Limpeza dos dados

Agora que possuimos os dois datasets que serão utilizados, um para treinar o modelo e o outro que queremos prever, vamos começar a tratar os dados para que a lingagem informal seja minimizada, e assim tentar melhorar a precisão do modelo. Também faremos a mesma limpeza em outros três datasets que utilizaremos para validação do modelo.

import pandas as pd 
import numpy as np
import re

# importando dataset de reviews
data = pd.read_csv('reviews_geral.csv', index_col=0)
data.head()

# importando dataset de tweets
tweets = pd.read_csv('dataset.csv', index_col=0)
tweets.head()

# importando datasets de validação
B2W = pd.read_csv('B2W-Reviews01.csv', delimiter=';')
tweets_aleatorios = pd.read_csv('NoThemeTweets.csv')
imdb = pd.read_csv('imdb-reviews-pt-br.csv')

B2W.head(1)

tweets_aleatorios.head(1)

imdb.head(1)

# função para trocar a nota 5 da avaliação por 1 (positivo) e a nota 1 por 0 (negativo)
def transforma_categoria(valor, categoria=1):
    if valor == categoria:
        return 0
    else:
        return 1

# função para trocar 'pos' por 1 (positivo) e 'neg' por 0 (negativo)
def transforma_categoria2(valor):
    if valor == 'neg':
        return 0
    else:
        return 1
    
# função para trocar 'Positivo' por 1 (positivo) e 'Negativo' por 0 (negativo)
def transforma_categoria3(valor):
    if valor == 'Negativo':
        return 0
    else:
        return 1

# aplicando a função de transformação da categoria
data['score'] = data.score.map(transforma_categoria)

imdb['sentiment'] = imdb.sentiment.map(transforma_categoria2)
tweets_aleatorios['sentiment'] = tweets_aleatorios.sentiment.map(transforma_categoria3)
#B2W['overall_rating'] = B2W.overall_rating.map(transforma_categoria)

# função para remover letras duplicadas
def removendo_duplicadas(texto):
    lista = []
    atual = ''
    anterior = ''
    for i in texto:
        if i != atual:
            lista.append(i)
            atual = i
        elif atual == 'r' or atual == 's':
            lista.append(i)
    return(''.join(lista))

#função para limpeza do texto
def limpa_texto(text):    
    text = str(text)
    text = re.sub('ã','a',text)
    text = re.sub('á','a',text)
    text = re.sub('à','a',text)
    text = re.sub('â','a',text)
    text = re.sub('ó','o',text)
    text = re.sub('õ','o',text)
    text = re.sub('ô','o',text)
    text = re.sub('é','e',text)
    text = re.sub('ê','e',text)
    text = re.sub('í','i',text)
    text = re.sub('ú','u',text)
    text = re.sub('ç','c',text)    
    text = re.sub(r'https?://\S+|www.\S+|@\S+', '', text)
    text = re.sub(r'<.*?>', '', text)
    text = re.sub(r'[^a-zA-Z0-9]+', ' ', text)
    text = re.sub(r'[0-9]', '', text)
    text = text.lower()
    text = removendo_duplicadas(text)
    return text

# aplicando a limpeza do texto nos datasets
data['content'] = data.content.map(limpa_texto)
tweets['full_text'] = tweets.full_text.map(limpa_texto)

#validacao
B2W['review_text'] = B2W.review_text.map(limpa_texto)
imdb['text_pt'] = imdb.text_pt.map(limpa_texto)
tweets_aleatorios['tweet_text'] = tweets_aleatorios.tweet_text.map(limpa_texto)

# removendo colunas que não serão utilizadas
tweets.drop(columns=['id_str','in_reply_to_status_id_str'], inplace=True)

B2W.drop(columns=['submission_date','reviewer_id','product_id','product_name','product_brand','site_category_lv1','site_category_lv2','review_title','recommend_to_a_friend','reviewer_birth_year','reviewer_gender','reviewer_state'], inplace=True)
imdb.drop((['text_en','id']), axis=1, inplace=True)
tweets_aleatorios.drop((['id','tweet_date','query_used']), axis=1, inplace=True)

# removendo vazios
tweets.drop(tweets[tweets['full_text'] == ''].index, inplace=True)
tweets.drop(tweets[tweets['full_text'] == ' '].index, inplace=True)
data.drop(data[data['content'] == ''].index, inplace=True)
data.drop(data[data['content'] == ' '].index, inplace=True)

# Validação
# removendo vazios e avaliações 2 a 4 do B2W
B2W.drop(B2W[B2W['review_text'] == ''].index, inplace=True)
B2W.drop(B2W[B2W['review_text'] == ' '].index, inplace=True)
B2W.drop(B2W[B2W['overall_rating'] == 2].index, inplace=True)
B2W.drop(B2W[B2W['overall_rating'] == 3].index, inplace=True)
B2W.drop(B2W[B2W['overall_rating'] == 4].index, inplace=True)
# removendo vazios
imdb.drop(imdb[imdb['text_pt'] == ''].index, inplace=True)
imdb.drop(imdb[imdb['text_pt'] == ' '].index, inplace=True)
tweets_aleatorios.drop(tweets_aleatorios[tweets_aleatorios['tweet_text'] == ''].index, inplace=True)
tweets_aleatorios.drop(tweets_aleatorios[tweets_aleatorios['tweet_text'] == ' '].index, inplace=True)


data.score.value_counts()

imdb.sentiment.value_counts()

tweets_aleatorios.sentiment.value_counts()

# balanceando o dataset
tweets_aleatorios.drop(tweets_aleatorios[tweets_aleatorios['sentiment'] == 0][:256000].index, inplace=True)

tweets_aleatorios.sentiment.value_counts()

B2W.overall_rating.value_counts()

# balanceando o dataset
B2W.drop(B2W[B2W['overall_rating'] == 5][:20000].index, inplace=True)

# aplicando a função de transformação da categoria
B2W['overall_rating'] = B2W.overall_rating.map(transforma_categoria)

B2W.overall_rating.value_counts()

# verificando a existencia de nulos
tweets.full_text.isna().value_counts()

# verificando a existencia de nulos
data.content.isna().value_counts()

# verificando a existencia de nulos
imdb.text_pt.isna().value_counts()

# verificando a existencia de nulos
tweets_aleatorios.tweet_text.isna().value_counts()

# verificando a existencia de nulos
B2W.review_text.isna().value_counts()

# gravando dataset limpo
tweets.to_csv('dataset.csv')
data.to_csv('reviews_geral_limpo.csv')

# validação
imdb.to_csv('imdb_limpo.csv')
tweets_aleatorios.to_csv('tweets_aleatorios_limpo.csv')
B2W.to_csv('B2W_limpo.csv')

Para facilitar a aplicação dos modelos futuros, vamos deixar salvo um dataset de treino e teste.

from sklearn.model_selection import train_test_split

# separa dados em treino e teste
train, test, _, _ = train_test_split(data, data['content'], test_size=0.3, random_state=42)

# gravando um dataset de treino
train.to_csv('train.csv')

# gravando um dataset de teste
test.to_csv('test.csv')
