==> 05 - Analise de Sentimento - SVM.ipynb <==
## Etapa 5: Baseline com SVM

Para termos um baseline vamos utilizar um metodo de aprendizado supervisionado SVM. Treinaremos o modelo com o dataset do review e em seguida vamos prever os sentimentos no dataset dos tweets. 

%matplotlib inline

import pandas as pd
import numpy as np

from sklearn import svm

import sklearn
from sklearn import metrics

import warnings
warnings.filterwarnings('ignore')

#importa dataset
train = pd.read_csv('train.csv', index_col=0)
test = pd.read_csv('test.csv', index_col=0)

train.head(1)

# vetorizando os textos
from sklearn.feature_extraction.text import TfidfVectorizer

vect = TfidfVectorizer(ngram_range=(1,4), min_df=1, max_df=0.95, decode_error = 'ignore')
vect.fit(train.content.values.astype('U'))

text_vect = vect.transform(train.content.values.astype('U'))
text_vect_test = vect.transform(test.content.values.astype('U'))

# treinando o modelo
clf_svm = svm.SVC(random_state=100)
clf_svm.fit(text_vect, train.score)

# avaliando o modelo
y_pred = clf_svm.predict(text_vect_test)

print("Accuracy:", metrics.accuracy_score(test.score, y_pred))

Para nossa baseline com um SVM conseguimos uma acuracia de 99 %, agora vamos rotular o dataset dos tweets

dataset = pd.read_csv('dataset.csv', index_col=0)

dataset.head(1)

# transformando em vetor os tweets
tweets_vect = vect.transform(dataset.full_text.values.astype('U'))

# prevendo os tweets
y_pred = clf_svm.predict(tweets_vect)

# adicionando os valores preditos na colina sentimento
dataset['sentimento'] = y_pred

dataset.head()

dataset.sentimento.value_counts()

dataset.to_csv('dataset_com_sentimento.csv')

print('Arquivo Gravado com Sucesso')
