==> 07 - Análise de sentimentos - RNN.ipynb <==
## Etapa 7: Treinando o modelo com uma RNN

Por fim, faremos uma nova predição do mesmo, porém utilizando uma rede neural recorrente, normalmente mais indicadas para trabalhar com texto.

# importa funções padroes
import pandas as pd 
import numpy as np
import re

# importa funções para redes neurais
from keras.preprocessing.text import Tokenizer
from keras.preprocessing import sequence
from keras.models import Sequential
from keras.utils import to_categorical
from keras.layers import Dense, Dropout, Embedding, LSTM, Bidirectional

def predicao(lista_texto):
    texto_token = tokenizer.texts_to_sequences(lista_texto)
    texto_padded_tokens = sequence.pad_sequences(texto_token, maxlen=100)
    return (model.predict(texto_padded_tokens[:len(lista_texto)]) > .5).astype(int)
    #return model.predict(texto_padded_tokens[:len(lista_texto)]).argmax(1)

def predicao_pesos(lista_texto):
    texto_token = tokenizer.texts_to_sequences(lista_texto)
    texto_padded_tokens = sequence.pad_sequences(texto_token, maxlen=100)
    return model.predict(texto_padded_tokens[:len(lista_texto)])

#importa dataset
train = pd.read_csv('train.csv', index_col=0)
test = pd.read_csv('test.csv', index_col=0)

X_train = train['content'].values.astype(str)
X_test = test['content'].values.astype(str)
y_train = train['score']
y_test = test['score']

max_features = 40000
maxlen = 250

# treina um tokenizador nos dados de treino
tokenizer = Tokenizer(num_words=max_features)
tokenizer.fit_on_texts(X_train)
len(tokenizer.word_index)

# tokeniza os dados
X_train_token = tokenizer.texts_to_sequences(X_train)
X_test_token = tokenizer.texts_to_sequences(X_test)

# corta ou adiciona zeros em sequências maiores de 250
X_train_padded = sequence.pad_sequences(X_train_token, maxlen=maxlen)
X_test_padded = sequence.pad_sequences(X_test_token, maxlen=maxlen)
X_train_padded.shape, X_test_padded.shape

y_train = np.asarray(y_train).astype(np.float32)
y_test = np.asarray(y_test).astype(np.float32)

# monta o modelo de rede neural
model = Sequential()
model.add(Embedding(max_features, 64, input_length=maxlen))
model.add(Bidirectional(LSTM(32, return_sequences=True)))
model.add(Dropout(0.3))
model.add(Bidirectional(LSTM(32)))
model.add(Dropout(0.3))
model.add(Dense(1, activation='sigmoid'))

model.compile('adam', 'binary_crossentropy', metrics=['accuracy'])

model.summary()

# treina a rede neural
model.fit(x=X_train_padded, y=y_train, batch_size=256, epochs=7)

#comparar os resultados
acc = model.evaluate(x=X_test_padded, y=y_test)

# serialize model to JSON
model_json = model.to_json()
with open("model.json", "w") as json_file:
    json_file.write(model_json)
# serialize weights to HDF5
model.save_weights("model.h5")
print("Modelo salvo em disco!")

Rotulando o dataset dos Tweets com a previsão do modelo RNN

# importando dados dos tweets
dataset = pd.read_csv('dataset_com_sentimento.csv', index_col=0)

# realizando a predição na base dos tweets
y_pred = predicao(dataset.full_text)
y_pred_score = predicao_pesos(dataset.full_text)

dataset['sentimento_RNN'] = y_pred
dataset['pesos_RNN'] = y_pred_score

dataset.head()

dataset.to_csv('dataset_com_sentimento.csv')

print('Arquivo Gravado com Sucesso')
