==> 09 - Analise dos rótulos gerados.ipynb <==
## Etapa 9 : Analisando os Rótulos gerados

Nesta etapa, vamos analisar as diferenças de rotulagem e definir o rótulo que será usado para o modelo final

# importa funções padroes
import pandas as pd 
import numpy as np
import re

#importa dataset
tweets = pd.read_csv('dataset_com_sentimento.csv', index_col=0)
tweets

tweets[(tweets["sentimento"] != tweets["sentimento_CNN"]) | (tweets["sentimento"] != tweets["sentimento_RNN"])]

Aproximada 30% dos tweets tiveram alguma divergencia entre os 3 modelos.

len(tweets[(tweets["sentimento"] != tweets["sentimento_CNN"])])

len(tweets[(tweets["sentimento"] != tweets["sentimento_RNN"])])

len(tweets[(tweets["sentimento_RNN"] != tweets["sentimento_CNN"])])

Em aproximadamente 23% dos tweets existe divergencia entre a predição da SVM contra a predição de algum modelo de deep learning, ja a diferença entre os modelos de deep learning cai para 15%. 

len(tweets[(tweets["sentimento"] != tweets["sentimento_CNN"]) & ((tweets["pesos_RNN"] > 0.9) | (tweets["pesos_RNN"] < 0.1))])

len(tweets[(tweets["sentimento"] != tweets["sentimento_RNN"]) & ((tweets["pesos_RNN"] > 0.9) | (tweets["pesos_RNN"] < 0.1))])

len(tweets[(tweets["sentimento_RNN"] != tweets["sentimento_CNN"]) & ((tweets["pesos_RNN"] > 0.9) | (tweets["pesos_RNN"] < 0.1))])

Este valor cai quase pela metade se avaliarmos apenas os tweets com um grau alto de certeza.

Para treinar nosso modelo final, vamos utilizar a mediana dos rotulos dos 3 modelos.

colunas = tweets.loc[: , "sentimento":"sentimento_RNN"]

media = []
for i in colunas.mean(axis=1):
    if i > 0.5:
        media.append(1)
    else:
        media.append(0)

tweets['sentimento_medio'] = media

tweets.head(20)

tweets.to_csv('dataset_com_sentimento.csv')

print('Arquivo Gravado com Sucesso')
