==> 12 - Analisando os Resultados.ipynb <==
## Etapa 12: Analisando os resultados

import pandas as pd 

def destaca_maior(s):
    is_max = s == s.max()
    return ['background-color: #99FF99' if v else '' for v in is_max]

def destaca_menor(s):
    is_min = s == s.min()
    return ['background-color: #99FF99' if v else '' for v in is_min]

## Modelos treinados com o dataset dos reviews

resultados = {'Modelo': ['SVM', 'CNN', 'RNN'],
              'Tempo Total de Treinamento': [260, 75, 224],
              'Tempo por Epoca': ['NA', 5, 32],
              'Nr de Epocas': ['NA', 15, 7],
              'Acc Treino': [99.25, 97.43, 97.87],
              'Acc Teste': [94.87, 95.35, 94.86]}

df = pd.DataFrame(resultados)
df = df.set_index("Modelo")
df.style.apply(destaca_maior, subset=['Acc Teste']).apply(destaca_menor, subset=['Tempo Total de Treinamento'])

Neste estudo a Rede Neural Convolucional se mostrou mais performática, tanto em tempo de treinamento quanto em acurácia, apesar de que podemos considerar um empate tecnico neste último quesito.

## Descrição:

**SVM** - Support Vector Machine   
**CNN** - Convolutional Neural Network   
**RNN** - Recurrent Neural Network   

Os tempos estão em minutos.  
Todos os modelos foram treinados em um notebook Dell, processador i5 7200 2.5Gh, 8Gb de memória RAM e um SSD 480Gb.

## Modelo Final

Para a escolha do modelo final, foram testadas algumas variações em termo de camadas, numeros de epocas e tamanho da rede. Além de dois datasets diferentes, no modelo 1 (**M1**), é utilizado apenas o dataset dos tweets rotulados(**Simples**), já no modelo 2 (**M2**), o dataset dos tweets rotulados é concatenado com o dataset dos reviews dos app na Gogle Store(**Agrupado**).


modelo_final = {'Modelo':           ['M1','M1','M1','M1','M1','M1','M1','M2','M2','M2','M2','M2','M2'],
                'Tempo por Epoca':  [420,460,183,183,220,220,250,2580,2827,980,980,1139,1139],
                'Nr de Epocas':     [5,5,5,10,5,10,5,5,5,5,10,5,10],
                'Acc Treino':       [96.70,96.76,51.51,87.70,90.73,93.47,53.07,97.34,97.36,92.65,94.73,91.39,95.31],
                'Loss Treino':      [0.0932,0.0899,0.6990,0.3660,0.3092,0.2434,0.6688,0.0820,0.0790,0.2472,0.2009,0.2690,0.1761], 
                'Acc Teste':        [88.39,88.01,51.88,84.46,85.86,85.79,52.26,94.07,94.16,86.54,92.58,87.67,93.16],
                'Loss Teste':       [0.3284,0.3487,0.6809,0.4245,0.4169,0.4336,0.6674,0.1880,0.1827,0.3980,0.2578,0.3249,0.2238],
                'Nr de camadas':    [2,2,1,1,1,1,1,2,2,1,1,1,1],
                'Dataset':          ['Simples','Simples','Simples','Simples','Simples','Simples','Simples','Agrupado','Agrupado','Agrupado','Agrupado','Agrupado','Agrupado'],
                'RNN':              ['LSTM','LSTM','LSTM','LSTM','LSTM','LSTM','GRU','LSTM','LSTM','LSTM','LSTM','LSTM','LSTM'],
                'Tamanho da rede':  [64,128,64,64,128,128,128,64,128,64,64,128,128]}

df_f = pd.DataFrame(modelo_final)
df_f.set_index("Modelo")

Observação: *Os tempos estão em segundos*

Definimos a escolha dos modelos pela melhora acurácia nos testes, por isso o **M1** ficou sendo com 2 camadas LSTM e tamanho 64 e o **M2** ficou com 2 camadas LSTM e tamanho 128. 

Comparando os dataset de validação

validacao = {'Dataset': ['IMDB','Tweets Aleatórios','Reviews Americanas','Tweets Coletados','Ponto Frio','Netflix','Nubank'],
                'Acc MR': [65.65, 52.82, 91.01,'NA','NA','NA','NA'],
                'Acc M1': [50.86, 49.54, 59.92, 56.62, 58.56, 52.67, 56.53],
                'Acc M2': [69.86, 53.60, 92.50, 90.57, 87.60, 87.86, 84.99]}

df_f = pd.DataFrame(validacao)
df_f.set_index("Dataset")

## Descrição:

**Acc MR** - Acurácia do Modelo Treinado com o dataset dos reviews de app do Google Store utilizado para rotular o dataset dos tweets   
**Acc M1** - Acurácia do Modelo Treinado com o dataset dos tweets coletados que foram rotulados pelo MR  
**Acc M2** - Acurácia do Modelo Treinado com o dataset dos tweets coletados + o dataset dos reviews de app do Google Store
